#0: Move ttnn::Shape and compute_strides to tt_metal - Modify compute_strides logic for handling 0 volume shape
