+++ tt_metal/common/CMakeLists.txt
@@ -8,6 +8,7 @@ target_sources(
         core_coord.cpp
         mesh_coord.cpp
         metal_soc_descriptor.cpp
+        shape.cpp
         shape2d.cpp
         shape_base.cpp
         tt_backend_api_types.cpp
